Refactor Analyzer to workaround picking issue
archeogit internally uses pygit2 which, it turns out is not pickable. As
a result, analyzing multiple curated vulnerabilities in parallel (using
multiprocessing) is impossible because the arguments to `run_analysis`
must be pickable. The refactor implements a workaround by instantiating
instances of `Archeogit` in each process avoiding the need to pickle.